requirements.txt
-f /usr/share/pip-wheels
cairocffi==0.8.0
CairoSVG==2.1.2
cffi==1.11.2
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.8
django-braces==1.12.0
django-crispy-forms==1.7.0
django-dbbackup==3.2.0
django-filter==1.1.0
django-import-export==0.6.1
django-js-asset==1.1.0
django-mptt==0.9.0
django-polymorphic==1.3
django-smuggler==0.9.0
django-tables2==1.16.0
django-treebeard==4.2.0
django-widget-tweaks==1.4.1
et-xmlfile==1.0.1
html5lib==1.0.1
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
pdfrw==0.4
Pillow==4.3.0
pycparser==2.18
Pyphen==0.9.4
pytz==2017.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
WeasyPrint==0.42
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
